Cloze questions

Creating cloze subquestions

Moodle.cloze_subquestion — Type

The data structure to hold a cloze subquestion.

struct cloze_subquestion

type::String

grade::UInt64

answers::Vector{Union{Tuple{Union{AbstractString, Number}, Number}, Tuple{Union{AbstractString, Number}, Number, Number}}}

The possible values of type are

MC: Multuple Choice;

MCV: Multiple Choice Vertical;

MCH: Multuple Choice Horizontal;

SA: Short Answer;

NM: Numerical;

MCVS: Multichoice Vertical Radio Buttons

The following line creates a cloze_subquestion of type SA, grade 1 with correct answer 2.

s1 = cloze_subquestion( "SA", 1, [(2,true)])

The following creates a multiple choice subquestion of grade 1 with correct answers 2, 4 and incorrect answers 3, 5.

s2 = cloze_subquestion( "MC", 1, [(2,true),(4,true),(3,false),(5,false)])

Creating a cloze question

Moodle.cloze_question — Type

The data structure to hold a cloze subquestion.

struct cloze_question

title::String

text::AbstractString

subquestions::Vector{Moodle.cloze_subquestion}

defgrade::Int64

penalty::Float64

tags::Vector{String}

The simplest way to create a cloze question is to create the subquestions first (see above) and then the question itself.

q = cloze_question( "Even numbers", "Mark the even numbers from the following list {{1}}. How many numbers did you mark? {{2}}", [ s2, s1 ] )

The line above combines the two subquestions s1 and s2 into a single close type question.

One may supply the optional arguments defgrade, penalty (see the Moodle system documentation) and tags; the latter is an array of strings specifying the tags that appear with the question on Moodle.