Multiple choice questions

Moodle.multiple_choice_question — Type

The data structure to hold a multiple choice type Moodle question

struct multiple_choice_question

title::AbstractString

text::AbstractString

answers::Vector{Tuple}

penalty::Float64

tags::Vector{String}

defgrade::Int64

single::Bool

shuffle::Bool

wrongmarkzero::Bool

There are several ways to create multiple choice questions. The following line creates a multiple choice question that asks the students to mark the even numbers among 0, -1, and 2.

q = multiple_choice_question( "Even numbers", "Which of the following numbers are even?", [(0,true),(-1,false),(2,true)])

The same question can be created by specifying separately the list of right answers and the list of wrong answers as follows.

q = multiple_choice_question( "Even numbers", "Which of the following numbers are even?", [0,2], [-1] )

Alternatively, one might define this question by specifying the list of options and defining a boolean function that can be applied to these options and decides whether a certain option is right or wrong.

q = multiple_choice_question( "Even numbers", "Which of the following numbers are even?", [0,2,-2], x -> x%2 == 0 )

Multiple choice questions can also be defined by using parameters. Consider, for example, the following line.

q = multiple_choice_question( "Remainder", "Mark the integers that give remainder $[[1]]$ modulo $[[2]]$.", (2,4,(0,2,4,6,8,10)), (x,y,z) -> z % y == x )

This line creates a question asking which of the numbers 0, 2, 4, 6, 8, 10 give remainder 2 modulo 4. The first two arguments are the name of the question and the question text, respectively. The third argument is a tuple whose first two entries are substituted in the placeholders [[1]] and [[2]], while the third argument is itself a tuple containing the options that appear among the possible answers. The last argument is a boolean function with three arguments. This function is applied to the first two entries of the tuple and to each entry of the third entry to calculate which of the answers are correct.

This parametric form can be used to create a large number of questions that correspond to a list of parameters. Take, for instance, the following example.

params = [ (a,b,(b,b+1,b+2,b+3,b+4)) for b in 10:20 for a in 1:4 ]
q = [ multiple_choice_question( "Divisible numbers", "Mark the integers that give remainder $[[1]]$ modulo $[[2]]$.", par, (x,y,z) -> z % y == x ) for par in params ]

In each of these versions, the function multiplechoicequestion has the following optional arguments:

penalty: see the Moodle system for documentation;

tags: array of strings specifying the tags that appear on Moodle after import;

defgrade: see the Moodle system for documentation;

single: true or false; if true then there is only one correct answer and the options appear on Moodle with radio buttons;

shuffle: true or false; controls if the answers should or should not be shuffled;

wrongmarkzero: true or false; controls if marking an incorrect answer should result in the reduction of the mark recieved by the student. The default is true and the marks for the answers are calculated in such a way that the students receive zero marks if they click all the options, and also when they leave all the options unclicked.