Short answer questions

Moodle.short_answer_question — Type

The data structure to hold a short answer type Moodle question.

struct short_answer_question

title::AbstractString

text::AbstractString

answer::Union{Number, String}

penalty::Float64

tags::Vector{String}

defgrade::Int64

One can use two methods to create a short anwser question. For example:

q1 = short_answer_question( "Question 1", L"What is the sum of $2$ and $-3$?", -1 );

This line creates a short answer question that asks for the sum of 2 and -3. Note that the question text is a LaTeXString.

The second way is to create questions with parameter. For example

q2 = short_answer_question( "Question 1", L"What is the remainder of $[[1]]$ modulo $[[2]]$?", (15,10), (x,y) -> x % y )

In the second case, the first and second arguments are the question title and the question string. The third argument is the question parameter. The two entries of this argument are going to be substtuted in the placeholders "[[1]]" and "[[2]]" in the question text. The fourth parameter is a function that has the same number of inputs as the number of entries in the question parameter; the function computes the right answer.

In the case of this simple example, the question q2 is the same as the question created by the following line.

q1 = short_answer_question( "Question 1", L"What is the remainder of 15 modulo 10?", 5 );

However, using the parametrised version, one can more easily create a larger number of questions.

params = [ (a,b) for b in 10:20 for a in b+1:2*b ];
qs = [ short_answer_question( "Question 1", L"What is the remainder of $[[1]]$ modulo $[[2]]$?", par, (x,y) -> x % y ) for par in params ];

These last two lines create 165 questions, one for each entry in the array params.

Both methods accept the optional arguments penalty, defgrade and tags. For the meaning of penalty and defgrade, consult the documentation of Moodle. The argument tags is a list of strings that will appear as tags for the question once imported from Moodle.

The second version also accepts the optional argument sep_left and/or sep_right that are strings marking the beginning and the end of the placeholders. The default values are "[[" and "]]", but these can be changed if the user needs these characters for other purposes.